ordered compositing
add alpha blend compositing to visit and use it to implement
ordered compositing. there is some logic to enable it automatically
and a gui option to explicitly disable it. it is on by default.
I also added an option for enabling depth peeling. this is off by
default because it's very slow. I changed all GetCell methods to
return long long so that overflow is less of an issue. I fixed a bug
in the transparency actor that was causing transparent geometry
to be sorted twice.


git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@27514 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Pipeline/Data/avtCellList.h
@@ -99,6 +99,9 @@ typedef struct
 //    Define private copy constructor and assignment operator to prevent
 //    accidental use of default, bitwise copy implementations.
 //
+//    Burlen Loring, Sun Sep  6 14:58:03 PDT 2015
+//    Changed the return type of GetNumberOfCells to long long
+//
 // ****************************************************************************
 
 class PIPELINE_API avtCellList
@@ -118,7 +121,7 @@ class PIPELINE_API avtCellList
     char                    *ConstructMessages(avtImagePartition *, char **,
                                                int *);
     void                     EstimateNumberOfSamplesPerScanline(int *);
-    int                      GetNumberOfCells(void) const { return celllistI; };
+    long long                GetNumberOfCells(void) const { return celllistI; };
 
     void                     Restrict(int, int, int, int);
 
@@ -126,8 +129,8 @@ class PIPELINE_API avtCellList
 
   protected:
     avtSerializedCell      **celllist;
-    int                      celllistI;
-    int                      celllistN;
+    long long                celllistI;
+    long long                celllistN;
     int                      nVars;
 
     bool                     useRestriction;
